Add update_market_data script
Scripts gathers tick data for markets specified in the 'allMarkets'
collection from the specified database and saves this data in
seperate collection for each of the markets in the same database.

At the moment script fails to gather data if there is too many markets.
Because it runs seperate thread for each market, pymongo and urllib
connection pools just go crazy throwing errno -3 when number of markets
exceed some limit.

TODO change the multithreading design to the thread pool design to
avoid errors when number of markets is too big.